+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Fixed
+- Bug which breaks sites using Akamai Bot Manager.
+
 ## 0.7.0 - 2026-06-26
 ### Changed
 - Use client spider option to not crawl new content.